refactor: repository, entity 패키지에 각 기능별 레포지토리와 엔티티 통합
