+++ src/main/java/com/dashboard/gmelan/config/roles/UserRole.java
@@ -1,0 +1,15 @@
+package com.dashboard.gmelan.config.roles;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+    ADMIN("ROLE_ADMIN"),
+    USER("ROLE_USER");
+
+    UserRole(String value) {
+        this.value = value;
+    }
+
+    private final String value;
+}
