+++ src/main/java/com/dashboard/gmelan/entity/Media.java
@@ -1,0 +1,39 @@
+package com.dashboard.gmelan.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.sql.Timestamp;
+
+@Entity
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name = "media", schema = "dashboard")
+public class Media {
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @Column(name = "id", columnDefinition = "BIGINT NOT NULL AUTO_INCREMENT")
+    private long id;
+
+    @Basic
+    @Column(name = "url", columnDefinition = "VARCHAR(50) NOT NULL")
+    private String url;
+
+    @Basic
+    @Column(name = "type", columnDefinition = "VARCHAR(10) NULL DEFAULT '분류 없음'")
+    private String type;
+
+    @Basic
+    @Column(name = "created_at", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private Timestamp createdAt;
+
+    @Basic
+    @Column(name = "updated_at", columnDefinition = "TIMESTAMP")
+    private Timestamp updatedAt;
+
+}
