+++ src/main/java/com/dashboard/gmelan/entity/Notice.java
@@ -1,0 +1,22 @@
+package com.dashboard.gmelan.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.sql.Timestamp;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Notice {
+    private long id;
+    private String title;
+    private Timestamp publishedAt;
+    private String content;
+    private String url;
+    private String agency;
+    private Timestamp fetchedAt;
+}
