+++ src/main/java/com/dashboard/gmelan/entity/ReferenceCategory.java
@@ -1,0 +1,39 @@
+package com.dashboard.gmelan.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.sql.Timestamp;
+import java.util.List;
+
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+@Table(name = "reference-category", schema = "dashboard")
+public class ReferenceCategory {
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @Column(name = "id", columnDefinition = "BIGINT NOT NULL AUTO_INCREMENT")
+    private long id;
+
+    @Basic
+    @Column(name = "name", columnDefinition = "VARCHAR(20) NOT NULL UNIQUE DEFAULT '기본 분류'")
+    private String name;
+
+    @OneToMany(mappedBy = "referenceCategory", cascade = CascadeType.ALL)
+    private List<Reference> references;
+
+    @Basic
+    @Column(name = "created_at", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
+    private Timestamp createdAt;
+
+    @Basic
+    @Column(name = "updated_at", columnDefinition = "TIMESTAMP")
+    private Timestamp updatedAt;
+
+}
