+++ src/main/java/com/dashboard/gmelan/entity/enums/UserPermission.java
@@ -1,0 +1,7 @@
+package com.dashboard.gmelan.entity.enums;
+
+public enum UserPermission {
+    USER,
+    ADMIN,
+    OPERATIOR;
+}
