--- src/main/java/com/dashboard/gmelan/user/config/UserRole.java
@@ -1,15 +1,0 @@
-package com.dashboard.gmelan.user.config;
-
-import lombok.Getter;
-
-@Getter
-public enum UserRole {
-    ADMIN("ROLE_ADMIN"),
-    USER("ROLE_USER");
-
-    UserRole(String value) {
-        this.value = value;
-    }
-
-    private final String value;
-}
