--- src/main/java/com/dashboard/gmelan/user/enums/UserPermission.java
@@ -1,7 +1,0 @@
-package com.dashboard.gmelan.user.enums;
-
-public enum UserPermission {
-    USER,
-    ADMIN,
-    OPERATIOR;
-}
