--- src/main/java/com/dashboard/gmelan/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.dashboard.gmelan.user.repository;
-
-import com.dashboard.gmelan.user.Entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    UserEntity findByUsername(String username);
-
-    UserEntity findById(long userId);
-
-}
