--- src/main/java/service/VelogService.java
@@ -1,11 +1,0 @@
-package service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class VelogService {
-    // TODO: https://api.velog.io/rss :최신 velog 포스트 목록 가져오기
-    // TODO: https://api.velog.io/rss/@{username} : 특정 사용자의 게시물 구독
-
-
-}
